Rename headers. Split User from client api header in its own file. Additional comments
